requirements.txt
certifi==2019.11.28
chardet==3.0.4
chardet2==2.0.3
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.7